BZ-614591: fix inventory sync issues that manifest during specific timings of async uninventory
part 1 - do not merge inventory report resources that have already been uninventoried

there exists a small window of time after the synchronous part of the uninventory and before the async
quartz job comes along to perform the actual removal of the resource from the database, that an inventory
report can come across the wire and !OVERWROTE! the UNINVENTORIED status back to COMMITTED.  if we find,
during an inventory report merge, that the existing resource was already uninventoried (indicating that
the quartz job has not yet come along to remove this resource from the database) we should stop all
processing from this node and return immediately.  this short-cuts the processing for the entire sub-tree
under this resource, but that's OK because the in-band uninventory logic will have marked entire sub-tree
for uninventory atomically.  in other words, all of the descendants under a resource would also be marked
for async uninventory too.

part 2 - async uninventory work should recursively delete resources

this handles the case when the the rare timings in part 1 made it into the system and the inventory is
current in an inconsistent state.  the likely effect is that a resource marked with UNINVENTORIED status
has a child that is marked as COMMITTED.  thus, if we recursively delete resources in the out-of-band job,
we ensure that synchronization inconsistencies are eliminated and the inventory is brought back to a
consistent state.

part 3 - async uninventory work should remove resources from their containing groups

even though the group removal occurs in the in-band work, there can be some group definitions that just
happens to perform its recalculation (either manually or schedules) in the period after the in-band work
completes but before the async job triggers. since the ExpressionEvaluator that underlies the bulk of the
dynagroup query generations automatically adds a filter to only manipulate COMMITTED resource, this work
should be a no-op most of the time.  however, in rare circumstances it's possible for an InventoryReport to
come across the wire and flip the status of resources from UNINVENTORIED back to COMMITTED.  in this case,
this group removal logic needs to be executed again just prior to removing the rest of the reosurce history.