BZ-615351: respect user's authorization when display suggestions
* only suggest resource-configuration property names for those resources which the logged in user has CONFIGURE permission on
* only return resource results whose configuration-specific search terms respect the aforementioned rule

-----

In order to make the query-translation routines aware of authorization,
the logged in user had to be passed to the concrete implementer of
AbstractSearchTranslator.  Since the translator is an implementation
detail hidden behind the CriteriaQueryGenerator, the logged in user thus
had to be passed to the generator as well.  In the future, when more
subsystems implement the search interfaces, they will automatically have
this additional contextual data available to them for advanced authz-based
filtering of suggestions and/or results.