BZ-616656: BZ-601189: fix searching and pagination on IE6
For IE6, revert back to a standard h:inputText element instead of the new
SearchBar (thus, IE6 will not support search suggestions or saved searches).
This eliminates the javascript error bubbling up from the SmartGWT library,
which prevents further interaction on the page from succeeding (clicking the go
button as well as pagination).