[BZ 573034] Updating script to not retrieve defaults from default template
The script originally fetched the default values for connection properties from the default
template and used the values from the default template to apply default values to connection
properties of resources. This is a nice solution because it is generic and applicable
across resource types; however, since 2.3.1 does not have the code changes in the plugin
upgrade code, the new properties do not exist in the default template in 2.3.1.
Consequently, the values are essentially hard-coded on a per resource type basis.