Cleaning up logic in script for resetting connection properties, adding script for apache
I have renamed reset_conn_props.js to reset_jboss_agent_prop.js. Most of the logic in
that script has been removed. There was logic for retrieving default values from the
default template; however, I discovered through further testing that the snapshot report
properties do not even exist in the default template when upgrading 2.2.0 --> 2.3.1.
Consequently, I need to simply hard code the properties that need to be added to the plugin
configurations.

Adding a similar script for apache to handle properties added in the 2.4.0 release.