Starting to implement criteria dsl.
The purpose of the dsl is to provide a more concise, yet readable syntax
for creating and executing criteria queries that are commonly used in the
remote apis and in the CLI. The syntax will look something like,

criteria(Resource) {
  filters = [
    id:           1
    name:         'My Resource'
    availability: UP
  ]

  sort = [
    name:       ASC,
    pluginName: DESC
  ]
}

This will create a ResourceCriteria object with the id, name, and
availability filters set. And it will sort by the name and pluginName
properties.