Adding ability for plugin jobs to persist data across invocations of jobs
Server plugin scheduled jobs previously had no mechanism provided by the plugin
framework for persisting data across invocations of the job. Quartz however
provides this capability when using a StatefulJob. Server plugin jobs run as
StatefulJob objects when they do not run concurrently. The only change needed
then was to expose a hook to the JobDataMap. A map has been added to
ScheduledJobInvocationContext which comes from the JobDataMap. Anything added
to this map by the server plugin job will be persisted from invocation of the
job to the next.

Initial commit for cloud plugin which at this point is just testing that
persisting job data works as expected.