Adding support for syncing the server endpoint url at start up.
The behavior is turned off by default. It can be enabled by setting

  rhq.sync.endpoint-address=true

in rhq-server.properties. When this is enabled, the server at start up will
compare its endpoint address to the host name/address found on the host machine.
If they differ, the server endpoint address will be updated to the value found
on the host machine. This addresses the (first part of the problem of)
address/host name changes that will occur in EC2 when a server machine is
restarted.