Restricting plugins to storing only string in the job data
Quartz is already configured to only allow strings to be stored in the job data
map when a job is scheduled for concurrent execution.  Restricting entries in
the map to strings is recommended to avoid class versioning errors that are
common during deserialization.  ScheduledInvocationContext has been refactored
a bit so that the underlying map is no longer directly exposed. To edit values
in the map it now exposes get/put methods that help enforce and ensure only
strings are stored in the map. There is still a getProperties() method; however,
it returns a read-only view of the data map.