Making return type of create/delete resource operations non-void
This change is primarily for remote clients like the CLI. If a remote client
creates or deletes a resource, it has no way of knowing whether or not the
operation succeeds since the methods in ResourceFactoryManagerRemote have a
return type of void. Now they return CreateResourceHistory and
DeleteResourceHistory objects so that the progress of those operations can be
tracked.