Only allow job data to be persisted when the job is not marked concurrent
Adding logic such that persisting job data in the
ScheduledJobInvocationContext is only supported when the job is not
marked as concurrent. If the job is marked as concurrent, and the plugin
calls a method to access or update properties, an exception will be
thrown.

Also adding some javadocs to explain how/when persisting job data is
supported