[BZ 619781] Handling more scenarios for settting connection properties to default values
Adding logic to handle three additional scenarios each of which involves
connection properties that are required and specify a default value:

1) A new property is added to a new group
2) An existing property not part of a group added to an existing group
3) A property previously in a group moved outside of a group

There were also some scenarios when properties were moved into or out of a group
and the group reference was not getting set or unset.
ConfigurationMetadataManagerBean has been modifed to update these references.