Fixing broken tests in UpdateConfigurationSubsystemTest
One test was failing due to a call to
PropertyDefinitionEnumeration.setOrderIndex in ConfigurationMetadataManagerBean.
That call should not be made because the owning PropertyDefinitionSimple object
takes care of calling setOrderIndex. The other test was failing due to a parse
error in a plugin descriptor that was using an attribute, 'default', that is no
longer supported for property options.