[BZ 614251] Making multiple changes so that live metric data can be fetched and displayed
There are two changes with this commit. First, in the CLI the ResourceProxy class
was using a criteria query that is too broad, at least when the jboss-as and
jboss-as-5 plugins are involved.

Secondly, the server/agent API used to get the
metric data is inadequate. It accepts a variable number of measurement names but
only a single data type for those measurements.  This is problematic because
there is nothing to guarantee that all of the metrics are of the same type.
On the server side, we were using a hard-coded value of METRIC for the data
type argument. The method(s) MeasurementManagerLocal.findLiveData and
MeasurementManagerRemote.findLiveData will fail any time one of the
measurement definitions has a data type other than METRIC.
MeasurementManager.getRealTimeMeasurementValue has been refactored to take a list
of MeasurementDataRequest objects. Each request specifies the metric name along
with its corresponding data type.