Changing order of tests execution to work around an OOM error
When running all tests in core/domain locally with the new unitils-based
test, I started seeing OutOfMemoryErrors with a message about the GC
running too much. I did not however, see this on hudson. I came to the
general conclusion that the problem stemmed from loading entities both
in the embedded ejb container and both in unitils which is done in
stand-alone hibernate as far as I know. I have updated the test suite so
that the unitils tests run after all of the tests that depend on the
embedded container, allowing the embedded container to shutdown.