Reject inventory reports that contain "stale" resource types
If an inventory report that is sent up to the server contains a resource
type that is marked for deletion, DiscoveryBossBean will throw an
exception letting the plugin container know that it needs to restart so
that it can update its plugins and resource types.

This is also the initial commit for the domain-test-utils module. I have
pulled ResourceBuilder and ResourceTypeBuilder out of the cli module and
moved them into this new module so that they can more easily be shared
across modules.