Reject incoming reports (on the server) that contain stale resource types
This commit contains two changes needed for BZ 549852. First, filtering
logic has been added in DiscoveryBossBean.mergeInventoryReport to reject
inventory reports that include any resource types that have been marked
for deletion. A report is rejected by throwing a StaleTypeException.

The second change is handling the StaleTypeException in the agent and in
the plugin container. A new listener interface, RebootRequestListener,
has been added to the PC. When InventoryManager catches a
StaleTypeException, it notifies the listener that a reboot is needed.
This listener is the agent. The agent shuts down the PC, clears out the
data directory, and then restarts the PC.