Removing circular dependency that was caused by client-api-test-utils
client-api-test-utils was introduced as a shared, test helper module
that is used both by client-api and by server/jar (so far). The problem
is that client-api-test-utils also depends on client-api, creating a
circular dependency. I have moved the single class in
client-api-test-utils into client-api. We are  now generating a test jar
that other modules can consume.