Clean up database after tests have run
The tests in ResourceMetadataManagerBeanTest have to commit changes to
the db and leaving plugins in the database in the RHQ_PLUGINS table is
causing subsequent failures in DatabaseAndFilePluginDeploymentTest. I
have added logic to remove all plugins added to the database during
test runs in ResourceMetadataManagerBeanTest.

Also adding some more tests for verifying behavior when resource types
are removed.