Add logic to delete bundles and bundle types
ResourceMetadataManagerBean lacks support for deleting content-related
stuff. With this commit, it now deletes bundles and bundle types that
are associated with the resource type being deleted. I have also added
tests to verify that bundles and bundle types are deleted.

I had to add some additional set up code to start the master (server)
plugin container. When I first wrote my test and a call to
BundleManagerBean, an exception was thrown complaining about the master
plugin contain not running.