Adding support for deleting alerts by template
When deleting a resource type, the alert definition templates for that
type need to be deleted as well. If they are not deleted, we run into a
FK constraint violation. This commit add JPQL queries and logic to
AlertManagerBean.deleteAlertsByContext to support deleting by template.

I have also updated and added tests in AlertManagerBeanTest. I have made
a significant change in the set up/tear down logic by utilizing dbunit.
By using dbunit we can more easily ensure that the relevant parts of the
db are in a known, consistent state before and after test runs. Tables
are seeded with data before each test, and then those tables are wiped
clean afterwards to avoid introducing side affects for other test
classes.