Initial commit for OperationMetadataManagerBean
The code for updating/deleting operation definitions in
ResourceMetadataManagerBean has been moved into
OperationMetadataManagerBean.

MetadataManagerTest now uses dbunit to clear the database prior to test
runs. The reason for this change is because after each test run of a
MetadataManagerTest (i.e., ContentMetadataManagerBeanTest) and prior to
its next run, the database has to be reset. Prior to this commit, this
has been accomplished by running dbsetup. That approach was too
inefficient. Now tests can be run without running dbsetup.

Lastly, fixing the annotations in MeasurementViewManagerTest. It was
using @BeforeTest and @AfterTest when I think the intent was
@BeforeMethod and @AfterMethod.