Initial commit for EventMetadataManagerBeanTest
Changing more @BeforeSuite annotations to @BeforeClass. This seems to
be causing failures on the delete-agent-plugin branch but not in master.
The failures make sense to me, but I do not understand 1) why they are
not also happening in master and 2) why I haven't seen them sooner since
the use of @BeforeSuite is not a recent change. With all that said, in
each case, the test class definitely should be using either @BeforeClass
or @BeforeMethod, definitely not @BeforeSuite.