Manual Import/Child Create work
Add support for manual import (aka manual add) and child resource creation
from the resource tree context menu (right click on a resource in the tree).
This required some new upload file support in order to upload the local
file content to be used for package backed resources. The support in place
for bundle distro file and bundle file support has been refactored to
share some code to support this and future upload needs. In particular, see
the new FileUploadForm base class supporting all of the current upload forms.

Creation of a package backed resource is now primarily done in two steps.
First, the backing package is uploaded and used to create a PackageVersion.
The PackageVersion is then used to back the resource in the resource create.
This approach is conducive to gwt not wanting to deal with java.io classes.
The new createResource signature may also be useful in the future for
selecting a backing package from a repo, for a new resource.

There are two new wizards, one for manual import and one for child
create. They are both built on AbstractResourceFactoryWizard.

ToDo: make sure Cancel cleans up and Previous/Next works well.


repo