Initial commit for PluginManagerBean and PluginManagerLocal
This is another major refactoring of ResourceMetadataManagerBean. There
was a significatn amount of code that dealt with the plugin as a single
unit as opposed to dealing with the plugin as a collection of meta data.
Examples of the former case include methods for enabling/disabling a
plugin and query methods for a Plugin. That code has been pulled out
into a new ejb, PluginManagerBean. I was very cognisant of the existing
transaction boundaries in ResourceMetadataManagerBean, I was careful to
maintain them.