Initial commit for InventoryManagerBean
InventoryManagerBean is an ejb that provides the API for performing the
in-band deletion of resource types. Initially I was going to put the
method in ResourceTypeManagerBean, but this would have created a
circular dependency between ResourceTypeManagerBean and
ResourceManagerBean, resulting in deployment errors. Unit/integration
tests will be coming in a subsequent commit.