Initial commit for PurgeResourceTypes, the scheduled job to remove resource types
PurgeResourceTypes is the scheduled job that carries out the out of band
work to permanently remove resource types that have been deleted. In
this context, deleted means that the resource type's deleted flag has
been set. I have added a few methods to InventoryManagerBean, in part
to minimize the dependencies for PurgeResourceTypes which will in turn
simplfy testing.

The first method added to InventoryManagerBean is getDeletedTypes which
just queries for deleted resource types. The second method is
isReadyForPermanentRemoval. This method encapsulates the logic for
determing if a resource type can be safely removed from the system. The
last method is purgeDeletedResourceType which does the removal. This
method is (currently) just a thin wrapper around
ResourceMetadataManagerBean.completeRemoveResourceType. A lot of the
work in completeRemoveResourceType is done via cascading deletes which
might not be the most performant. This code needs further analysis (and
refactoring).