Filter out resource types that have already been deleted from the db
DeletedResourceTypeFilter previously only filter out resource types that
were marked deleted. If a resource in an inventory report was of a type
that was already removed from the database, it would get past the
filter. Since the type did not exist on the server, it would not get
committed into inventory, but the server would just ignore it when it
should be instead notifying the agent that it sent up a stale inventory
report.

When the agent is notified that plugin container needs to be restarted,
it now updates its plugins with the server after shutting down the PC
and prior to starting it back up. This takes care of removing plugin
that were deleted on the server.