- Add Delete capability into the resource search views. The delete button is enabled only when the selected resource's types support delete and the user has delete permission on the selected resources.  To do this checking added ResourceCompositeSearchView and ResourceCompositeDataSource, which provide resource permission info at resource row level granularity.  Perf here could be improved by adding a a ResourceCompositeCriteria (extending ResourceCriteria) that could optionally pull composite fields. I'll probably add that shortly, and pull only the resourcePerms as additional data. -- Added deleteResources to the ResourceFactoryManagerRemote
- Add Import/Create New actions to the Inventory->Children subtab.  The user
selects the type and then the wizard is invoked. Working the type
selection into the wizards proved difficult given that the type actually
helps determine the wizard's steps.  So, instead, added the ability
to create TableActions that are menu buttons as opposed to just buttons
(unfortunately I ended up having to change the TableAction.executeAction()
signature so a bunch of files gor touched).
Import and Create are only rendered when the parent type has appropriate
child types supporting those aactions, and the user has permission on the
parent to perform those actions.