Updating plugin deletion logic for HA deployments
Initially I was deleting the plugin from the RHQ_PLUGIN tableas well as
deleting the plugin jar file from the file system. The problem with
deleting the plugin from the database in an HA deployment, as mazz
explained to me, is that servers other than the one that serviced the
request, will not know that the plugin has been deleted. If the agent
plugin scanner comes along and sees the plugin jar file on disk, it
would attempt to re-install the plugin.

AS far as the Plugin object/file goes, deletion of agent plugins now has
the same semantics as it does with server plugins in so far as it state
that the plugin assumes. When the agent plugin scanner runs, it now
looks at any plugins that have been deleted and removes them from the
file system.

A separate purge operation will be added to PluginManagerBean in a
subsequent commit.