Adding support for purging an agent plugin
I have updated the UI for agent plugins to behave similar to how it
behaves for server plugins. There is now a 'SHOW DELETED' button on the
agent plugins tab. When you click it, deleted plugins will be displayed
as well as a 'PURGE' button. Clicking the purge button will result in
the plugin being purged from the database.

The purge operation as a safeguard calls the delete operation on each
plugin before purging them.