Making transactions for meta data updates more granular to improve performance
After doing some profiling with jprofiler I have decided to make each
meta data manager update take place in its own transaction. We are
taking a big performance hit on all of the calls to
EntityManger.flush(). Doing less work per transaction takes away the
need to flush the persistence context. In addition to the changes in
transaction boundaries, I have removed several calls to flush(), most
notably in ConfigurationMetadataManagerBean. JProfiler flagged the code
in that EJB in particular as a hot spot.