Putting build number generation plugins in profile to facilitate building in mead
We execute git during the build to generate some build properties, namely a
build number. The git executable however is not available in the mead build
environment. I have move the plugins involved with generating build properties
into a profile so that we can disable them during builds in mead.

Conflicts:

	pom.xml