Adding profile to generate build number by specifying it manually
This is necessary for building in brew. We have to build RHQ in brew and
part of the build involves generating a build number which is used in
the app. We generate that build number by executing git, but the git
executable is not available in the brew/mead environment, so I am adding
a profile that takes the build number from the property rhq.build.number
which should be specified on the command line as a system property. This
will activate the profile. In conjunction with specifying that property,
the build-properties profile should be disabled with -P!build-properties