Add error handling logic to reboot listener and reboot agent instead of PC
This commit consists of a few changes. First, the reboot request
listener now reboots the agent instead of just the PC. Rebooting the
agent will rebot the PC, and it will also help minimize PermGen memory
leaks that we are susceptible to due to frequent class loader creation.

Secondly, the reboot request listener is now initialized prior to
initializing the PC. This will ensure we deal with any stale types that
might be encountered at start up. In order to support initializing the
listener first, it was moved out of InventoryManager and into the
PluginContainer itself. This has the added benefit of giving other parts
of the PC an easy way to request a reboot should the need arise.

Lastly, a substantial amount of error handling has been added to the
reboot listener. If shutting down the agent fails, we will try again
after a short delay. If the second attempt fails, we try to reboot just
the PC as a list ditch effort. If that fails, all we can do is report
the error indicating that a manual restart is needed. Once the agent is
shut down, we start it back up. If the start up fails, we trying
stopping/starting one last time. If the restart succeeds we increment
the agent restart counter.