Making StaleTypeException inherit from InvalidInventoryReportException
When I introduced StaleTypeException, I had it extend from
InvalidInventoryReportException. I had problems with it getting
wrapped as-is and sent down to the agent. I managed to resolve this by
making StaleTypeException instead extend from java.lang.Exception. The
reason my initial solution did not work is because I declared the
interface, DiscoveryServerService, to throw the exception, but I failed
to declare it in the implementation.