Remove query to fetch all resources of a particular type
The query to fetch all resources in
isReadyForPermanentRemoval(ResourceType) could be very expensive if
there are a lot of resources in the system. We only need to know if
there are any resources of that type, so I have removed the query in
place of one that just returns a count.