Refactoring filter to minimize the number of trips to the database.
Refactoring the logic in DeletedResourceTypeFilter for when a resource
type has already been purged from the database. Previously if the report
contained a type that was already deleted, I was querying the database
to see if the type's plugin existed. Now, the filter only hits the
database twice. Once to load deleted types, and once to load installed
plugins. The filter checks against installed plugins instead of against
deleted plugins because checking against deleted plugins would miss the
case in which the plugin has been purged from the database.