Adding queries to generate plugin stats
The stats include the number of resource types and resources of those
types in a given plugin. The idea is that these stats can be displayed
to the user immediately prior to deleting a plugin so that he is aware
of what is going to happen. These stats might otherwise be useful in
other areas of the app like a portlet that displays summary info on the
system.