[BZ 662068 all actions/buttons for recent alerts portlet always disabled. Unable to Ack alerts] In short this is due to the fact that the recent alerts portlet is like the recent alerts report, they are both "subsystem" views. As such the single alert listing can include alerts for which the user may or may not have alert (write) perms.  Both the portlet and the report view simply disallowed the buttons to be safe.  There are only really three alternatives, disallow all buttons, allow the buttons only for inv managers, or code up granular button enablement that involves perm checking on the resources backing each alert.
I took option 2 which at least allows inv managers the ability to process
the alerts en masse from the portlet or the report.  Other users will need
to navigate to the alert or resource from the links provided, and process
alerts from there.

This change makes available the user's global perms to any portlet, and
also any AbstractSectionedLeftNavigationView (the latter for the reports
view).

Also, fixed the alert links in the portlet. This required a slight
enhancement to TableSection, the ability to override the automatic html
link wrapping of "name" fields.