[BZ 669567] Don't update package types for a bundle plugin
Package types are not applicable to bundle plugins so we do not even
need to compare package type between the old and the new plugins. The
new bundle plugin will also have zero package types. The old plugin will
have pakcage types associated with it if there are deployed bundles.
This was resulting in our incorrectly deleting package types.