Selenium Locator Work - Fix a problem in AbstractSelector.onDestroy(). We have to be careful of destroying things manually that are also destroyed automatically by SmartGwt.  In general, destroying a parent will destroy it's children automatically. For example, destroying a SectionStack as a whole, destroying a Layout, Window, etc.
We are still usually on the hook for member canvas's.

Calling destroy() on a widget that has already been destroyed will
force an unwanted reinitialization (onInit) of the widget (just so it can
be properly destroyed immediately following). This behavior isn't great in
smartgwt, I wonder if they could change this...

This instance of the issue was in the Alert Notification Sender definition
call chain. I'm sure it would have cropped up in other places as well.

Also, in general we don't use markForDestroy() but rather destroy() for
immediate removal from the DOM. This protectes against selenium ID
conflict.

- there is still one onInit() reinvocation I'm not able to explain. I
protected against it.