Selenium Locator Work - Activity2 subtab
This class suffered from two main issues with its locator impl.
1) You don't need locators on "inaccessible" widgets.  Often H/V/Layouts
are used purely for formatting and are not really part of any widget
parent-child identification scheme. In this case just use non-Locatable
widgets to keep things as simple as possible. I made these conversions in
a few places.

2) You can't use static locatorIds on multiple widgets. Basically, avoid
re-using the same locatorId for widgets created inside of a loop. There
were a few places like this where result rows are being looped through.
Note that not all data is conducive to unique identifiers. In several places
the best that we can do is to use unique rowNums in the locators.

In general tried to shorten up he locator id strings as possible. They need
to be unique and descriptive, but as short as needed is good.