Updated test to be consistent with new role creation logic
Now when a role is created, subjects are associated with the role are
assigned the role. This assignment had to happen separately previously.
I also refactored the assert code to use assertXXX() methods instead of
the assert statement because the assertXXX() methods provide error
reporting out of the box that you do not get with the assert statement.