Adding the integration.ejb3 group back to the startEmbeddedJboss method
Trying to resolve test failures on hudson which I think is due to the
order of test execution. I see that tests in the plugin.metadata group
are executing before the startEmbeddedJboss method is called which
causes data source JNDI look ups to fail. The plugin.metadata group
depends on the integration.ejb3 group to ensure that the embedded
container is bootstrapped.