[BZ 677349] Do not obtain PC read lock prematurely
EventContextImpl was previously obtaining a PluginContainer read lock
prematurely so that it could obtain a reference to EventManager. I say
it was premature because there is no need to obtain the reference to
EventManager in the constructor. Its not needed until the different
instance methods are called and that is when the lock should be
obtained.

With this change when a plugin component tries to register event pollers
in thread T2, T2 will wind up blocking until the PC running in thread T1
has finished initializing and the PC in T1 releases its write lock.