fix: Bug 639950 - Resource tree-when user moves from one auto-group node                   to another, it doesn't navigate to the same tab/subtab.
This was a subtle problem in javascript vs java regex handling. This worked
in hosted mode (java eval) but not in standard mode (javascript eval). We're
unlikely to hit this again but in short watch out for disjunctives where
option 1 is a subset of a later option.

In Javascript:

"a/b/c".matches( "(a|a/b)/[^/]*" ) -> false

"a/b/c".matches( "(a/b|a)/[^/]*" ) -> true

In Java both of these resolve to true.  But in JS, the first pattern allows
JS to match on the "a" option and then try to match the rest, which fails. It
never goes back to realize that it could also match the "a/b" option and
continue on to a successful match of the whole string.