fix: Bug 565852 - Ugly error message when trying to "Ignore" a platform
The use case for ignoring a platform is a little fuzzy, why run the agent
in the first place.  But regardless, the root problem is that you get a
bad error trying to ignore a server resource beneath an uncommitted parent
platform.  In general the button enablement for AD was lacking so the fix
is to prevent invalid button enablement.

Also, allow a user to just select/import a platform without automatically
selecting the children.