Lazy Create SubTab Views This means that for resource and group detail views we no longer create the views for all active subtabs in advance. Instead, we create them when they are navigated to. This means that construction/init/oninit/destroy/onDestroy logis is avoided for non-rendered tabs.  This should make things a bit snappier overall and also protects against unwanted db round trips initiated (perhaps unintentionally) from onInit (or even construction).
In short, supply SubTabs with ViewFactories as opposed to the view itself,
and generate the Canvas as needed. Note that once created we re-use the canvas
until the entire detail view is destroyed.

- also, move async db loads from construction to oninit in resource
ActivityView and group Activity2View. In general avoid async work
and db rts in the constructor. Wait for the widget to be attached to a
parent (oninit), or even better, when rendered (onDraw).