[BZ 683859] Remove parent link to deleted reference is not accessible by parent
A user ran into the following situation in the CLI. He deleted a
resource  via ResourceFactoryManagerBean.deleteResource. He then issued
a criteria query to fetch the parent of the deleted resource. The
criteria was set to fetch child resources. The deleted resource was
included in the returned chiled resources even though it should not
have.

This commit removes the resource's parentResource reference when the
resource is deleted so that the deleted resource is not accessible by
parent.