Squahshing 13 commits from ec2 branch first commit message: Adding new operation to rhq-agent plugin to update the server endpoint url
Invoking the switchToServer operation will result in the operation switching
over to the specified server. In an EC2 deployment, the server can invoke this
operation whenever its IP address has changed. Then the agent will have a valid
server endpoint url from which it can download its failover list.

second commit message:
Initial commit for rhq-agent-wrapper-ec2 script

This is a wrapper script for rhq-agent-wrapper.sh that additionally configures
the server endpoint url and will also set up the agent to run as a service.

third commit message:
Adding logic to check for and skip initialization if it has already run

The initialization function only needs to run the first time an agent
runs so that the server endpoint url can be specified. That url will be
supplied in the AMI's user-defined data. Also adding logic to force
initialization to run when specifying the <init> cmd line arg.

fourth commit message:
Adding ability for plugin jobs to persist data across invocations of jobs

Server plugin scheduled jobs previously had no mechanism provided by the plugin
framework for persisting data across invocations of the job. Quartz however
provides this capability when using a StatefulJob. Server plugin jobs run as
StatefulJob objects when they do not run concurrently. The only change needed
then was to expose a hook to the JobDataMap. A map has been added to
ScheduledJobInvocationContext which comes from the JobDataMap. Anything added
to this map by the server plugin job will be persisted from invocation of the
job to the next.

Initial commit for cloud plugin which at this point is just testing that
persisting job data works as expected.

fifth commit message:
Adding support for syncing the server endpoint url at start up.

The behavior is turned off by default. It can be enabled by setting

  rhq.sync.endpoint-address=true

in rhq-server.properties. When this is enabled, the server at start up will
compare its endpoint address to the host name/address found on the host machine.
If they differ, the server endpoint address will be updated to the value found
on the host machine. This addresses the (first part of the problem of)
address/host name changes that will occur in EC2 when a server machine is
restarted.

sixth commit message:
Restricting plugins to storing only string in the job data

Quartz is already configured to only allow strings to be stored in the job data
map when a job is scheduled for concurrent execution.  Restricting entries in
the map to strings is recommended to avoid class versioning errors that are
common during deserialization.  ScheduledInvocationContext has been refactored
a bit so that the underlying map is no longer directly exposed. To edit values
in the map it now exposes get/put methods that help enforce and ensure only
strings are stored in the map. There is still a getProperties() method; however,
it returns a read-only view of the data map.

seventh commit message:
Updating plugin descriptor to make cloud plugin clustered

Also refactoring plugin to work with new methods in
ScheduledJobInvocationContext

eigth commit message:
Only allow job data to be persisted when the job is not marked concurrent

Adding logic such that persisting job data in the
ScheduledJobInvocationContext is only supported when the job is not
marked as concurrent. If the job is marked as concurrent, and the plugin
calls a method to access or update properties, an exception will be
thrown.

Also adding some javadocs to explain how/when persisting job data is
supported

ninth commmit message:
Fixing methods to throw the correct exception type

tenth commit message:
Adding logic to detect changed addresses

eleventh commit message:
Schedule the 'switchToServer' agent resource operation

twelth commit message:
Exposing address sync job as an operation that can be manually invoked.

This is a first code at this plugin operation. Some error handling needs
to be added along with some logging.

thirteenth commit message:
Adding logic to purge stale servers from the sync list

If a server has been taken down and removed from the system, then we
need to remove it from the list of servers we look at to see if their
addresses need to be synced.

Adding a bunch of debug logging.

Conflicts:

	modules/plugins/rhq-agent/src/main/resources/META-INF/rhq-plugin.xml