[BZ 677349] Adding logic to avoid deadlock during PC initialization
This commit adds an InitializationListener api to PluginContainer. When
plugin components start up while the PC is still initializing and
register event pollers, they now do so by way of this
InitializationListener api. EventContextImpl now registers a listener,
which does not require a lock and therefore eliminates the possibility
for deadlock. When the listener is invoked, the PC has already
released the write lock, EventManager is already up and running, and
there should be no problem registering the event poller at this point.
Also note the listener runs in the same thread that executes
PluginContainer.initialize.