[BZ 683545] scan plugins directory after deleting plugin and fix purge logic
This commit addresses a problem that occurs when deleting, purging, then
re-adding the plugin. The agent plugin scanner maintains a cache of
plugins on the file system, and that cache was getting dirty. This
results in the scanner failing to detect a plugin that has been re-added
as a new plugin, and consequently, the plugin did not get re-installed.
Now when deleting a plugin, a scan is immediately triggered.

Then if you went yo purge right away, you could wind up with a
constraint violoation in rhq_resource_type. Resource types are purged in
a scheduled job, so it was possible to purge and re-add a plugin before
its resource types have been purged. Now the plugin is marked to be
purged. The purge is done in a scheduled job and is not done until all
resource types from the plugin have been purged.

Once the user sees that the plugin has been purged, he knows it is safe
to re-add it.