in our GWT UI, the alert and event severity filters are represented with a drop down selection menu containing one checkbox for each severity enum (alert: hi, med, lo; event: debug, info, warn, et...). If you don't select any severity (i.e. all the checkboxes are unchecked), we used to assume this means there is no filtering to be performed and the query returns everything.
However, because the UI renders these in individual checkboxes, that behavior doesn't really make sense and will
confuse the user. Because, if the user doesn't select any severities, it
seems as though the user is asking it to show nothing - since no event or
alert will match the selected severities (and there are NO selected severities).
So to make the query results/UI more sensical, if the user de-selects all severity
check boxes, the table will show 0 rows.