A fix to eliminate duplicate refresh of table data (commit abc7cd665aa8a) exposed issues in some of our view impls that unintentionally relied on the refresh (second fetch) to work.
In short, the broken views need to supply initialCriteria for the first
fetch (nitiated by autoFetch=true on the listgrid).

But, we may not want that initialCriteria to be "fixed" for future fetches,
so I added an easy way to override that behavior.

This commit fixes the event history view. Subsequent commits will fix the
other broken views.