[BZ 691773] Make sure agent scanner always clears the scanned jar files
In registerAgentPlugins(), if an exception is thrown in the call to
ProductPluginDeployer.pluginDetected(), the exception propagates all the
way up the call stack, and AgentPluginScanner.scanned does not get
cleared. The call to scanned.clear() is now in a finally block to ensure
it is called even when an exception is thrown in ProductPluginDeployer.