[BZ 697610 - plugin and resource group config views still use old Resource disambiguation mechanism]
Added ancestry based disambiguation to the necessary config editing. It's
a bit different than the typical ListGrid approach as this is not a datasource
based table but rather a form.  So, instead, grab the display values via an SLSB
call.  No hover is available, if we feel we need that we may be able to
add it later.
- Added new EXPANDED Ancestry format, which prepends the actual resource
name.