[BZ  695753 - Cannot view traits for any resource on monitoring tab]
This occured only for non-permissioned users.  There were two problems
with the authz token handling.  First, the authz fragment needed to be be
overriden (in SLSB) as the "resource" entity must be accessed through
the "schedule" entity on MeasurementData.  Second, we should not be using
the GROUP authz token type regardless of the criteria, because MeasurementData
is at the resource level only. Meaning, we are not querying for "group-level"
data regardless of whether we are filtering on the group embership for the
target resource.