[BZ 697484 - Reports->Configuration History: Sorting on Ancestry column throws exception]
So actually, we're throwing exceptions all over the place on various column
sorting.  This check-in fixes, I think, all columns for all of the report
views.  Since several of those views use datasources common to other views
many of these fixes will apply in various places. So that is good.

But, I'm sure there are a lot of other places where we will still get hit.

The trick is that it's not always obvious that there is a problem because
if all of the data is held client-side (i.e. gwt has fetched all of the rows
and is holding them client-side then all rows will sort properly.  It's only
when sorting is applied server-side that things blow up.

So, when testing you must first visit the view. Then, select a column for sorting
and *then* hit the refresh button.  This will force a server-side fetch with
the sort applied.

Note that even setting ListGridField.setCanSortClientSideOnly(true) is not
solving the problem if applied to a field (although I think it should).

In certain places I've added the necessary fields to as Criteria sort fields
but this is not appropriate for every column.  In other circumstances you
can now perform the column-to-sort-field mapping in the datasource by
overriding the new "getSortFieldForColumn(String columnName)" method. And
by returning null you can ensure inappropriate fields do not make it to the
server for sorting.

- I may take out the calls to setCanSortClientOnly() as I'm not sure it buys
us anything.
- Also, I removed unused datasource field definition in a few places.